Swap out `github-markdown` for `redcarpet`
`github-markdown` seems to be deprecated and unmaintained, whereas
Redcarpet seems to be under active maintenance (and potentially faster
too).

This PR also removes support for the syntax highlighting filter - we were generating the markup but never styling it (except the outer `pre` tag) so continuing to process it in the pipeline was totally superfluous.

We will need to test this on a staging environment to make sure that nothing else is looking wacky, but the tests do all seem to be passing.

[#139711277]